25 Nov 2024
In class I did the main operating part of my code where it takes in the frequency and so far I feel like I'm imporivng on my logic for the program as a proceed to code each time. One challnge that I'm facing is that since the code doesnt actually get input of sound frequency it needs to read it from a set of data that I give to the code but then I couldnt understand how I would ask my program to read the data each second but alter on in class by doing reseacrh i found out the solution to my problem and now I designe dit so that:  processes sound data to detect meaningful baby crying patterns based on thresholds for frequency, duration, and pauses. It evaluates each sound sample, incrementing cry duration when the frequency exceeds 500 Hz, and resets it if pauses exceed 5 seconds, ultimately determining whether the baby needs attention.