28/11/24
I finalized my code by adding a randomizer to create inputs for sound frequency and cry duration so that the program can get inout as if program is inside an actual baby monitor.  Also my research stated that if the baby is actually awake and needs attention they usually cry between 2000-4000 Hz so I change the limit boundaries and I set the random generator to genera a number for hertz between 1-4000 hertz.